# Conflicts:
#	Cargo.lock
#	Cargo.toml
#	config/default.toml
#	config/tests.toml
#	docker-compose-develop.yml
#	src/application/db.rs
#	src/application/logger.rs
#	src/application/settings.rs
#	src/handlers/user/mod.rs
#	src/repositories/article/mod.rs
#	src/repositories/user/mod.rs

+++ src/domains/mod.rs
@@ -6,12 +6,12 @@ pub mod user;
 #[derive(Deserialize)]
 pub struct CollectionParams {
     #[serde(default = "default_per_page")]
-    pub per_page: usize,
+    pub per_page: i32,
     #[serde(default)]
     pub order: Option<String>,
-    pub last_id: Option<usize>,
+    pub last_id: Option<i32>,
 }
 
-fn default_per_page() -> usize {
+fn default_per_page() -> i32 {
     10
 }
